Move utility function from _utils.py to unifying.py
